\name{establishCleanSlate}
\alias{establishCleanSlate}
\title{establishCleanSlate}
\description{
Establishes clean slate.
}
\usage{
establishCleanSlate()
}
\arguments{
}
\value{
  A character vector of names of objects in current R global environment.
}
\examples{
establishCleanSlate()
cleanUp()

## Alternatively:
# cleanSlate <- establishCleanSlate()
# cleanUp(cleanSlate)
}
