\name{getRasterCoords}
\alias{getRasterCoords}
\title{getRasterCoords}
\description{
Outputs coordinates as complex numbers for a given image raster.
}
\usage{
getRasterCoords( raster, realY = FALSE )
}
\arguments{
\item{ raster }{ 2D numeric matrix / image. }
\item{ realY }{ Boolean. If \code{TRUE}, returns Y coordinate as the real component. }
}
\value{
  A matrix of complex numbers that represent x and y coordinates.
}
\examples{}
