\name{strsplit2}
\alias{strsplit2}
\title{strsplit2}
\description{
Splits a character vector using strsplit, and consolidates results as a matrix. Taken from \code{limma} package.
}
\usage{
strsplit2(x, split, ...)
}
\arguments{
\item{ x }{ Character vector. }
\item{ split }{ Character for splitting. See \code{strsplit}. }
\item{ ... }{  Other arguments passed to \code{strsplit}. }
}
\value{
  A matrix of split character vectors.
}
\seealso{
 \link{limma::strsplit2}
}
\examples{}
