\name{subsetDF}
\alias{subsetDF}
\title{subsetDF}
\description{
Subsets a dataframe columns starting with a given character.
}
\usage{
subsetDF(
    spotcalldf,
    prefix
)
}
\arguments{
\item{ spotcalldf }{ Dataframe containing per-pixel metrics. }
\item{ prefix }{ Character. Column prefix to subset by. }
}
\value{
  A matrix obtained from subsetting dataframe so column names all begin with \code{prefix}.
}
\details{
}
\seealso{
}
\examples{}
