\name{spotcall_troubleshootPlots}
\alias{spotcall_troubleshootPlots}
\title{spotcall_troubleshootPlots}
\description{
Creates plots for troubleshooting spotcalling pipeline.
}
\usage{
spotcall_troubleshootPlots(
    spotcalldf,
    imList,
    chosenCoordinate,
    plotWindowRadius = 10,
    decodeMetric = 'DECODE',
    params = get('params', envir = globalenv())
)
}
\arguments{
\item{ spotcalldf }{ Dataframe containing per-pixel metrics. }
\item{ imList }{ Named list of 2D numeric image matrices to register. Names should be \code{bit_names}, that indicate image and channel. }
\item{ chosenCoordinate }{ Complex or complex vector. Indicates coordinate central pixel(s) of window to plot.  }
\item{ plotWindowRadius }{ Integer. Number of pixels from central pixel to window parameter. }
\item{ decodeMetric }{ Character. References an object in the \code{imMetrics} environment (if \code{imMetrics} still exists). }
\item{ params }{ The parameters environment. }
}
\value{
  A list of plots.
}
\details{
  For each coordinate specified in \code{chosenCoordinate}, a plot is generated showing pixel intensity values across bits, with gene identities (determined by \code{spotcalldf}) overlaid atop the images. If  \code{chosenCoordinate} is not specified, defaults to the first pixel in \code{spotcalldf}.

  A list of images (\code{imList}) can be provided. Otherwise, will look for the object referenced by \code{decodeMetric} in \code{imMetrics} (generated by \code{getImageMetrics}). If \code{imMetrics} does not exist (and \code{imList} not provided), an error message is returned.

  Designed to work with \code{filterDF}.
}
\seealso{
\link{filterDF}
\link{getImageMetrics}
}
\examples{}
